% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barprob.R
\name{bar_prob}
\alias{bar_prob}
\title{Probability represented by a single bar}
\usage{
bar_prob(categories, probabilities, colours = NULL)
}
\arguments{
\item{categories}{Vector of category names}

\item{probabilities}{Vector of probabilities}

\item{colours}{Colours of the categories}
}
\value{
GGPlot object
}
\description{
Probability represented by a single bar
}
\examples{

bar_prob(c("Yes", "No"), c(80,20),
        colours=c("green", "red"))
bar_prob(c("Yes", "Maybe", "No"), c(50,30,20),
        colours=c("green", "orange", "red"))

}
