% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barprob.R
\name{barprob}
\alias{barprob}
\title{Probability represented by a single bar}
\usage{
barprob(categories, probabilities, colours = NULL)
}
\arguments{
\item{categories}{Vector of category names}

\item{probabilities}{Vector of probabilities}

\item{colours}{Colours of the categories}
}
\value{
GGPlot object
}
\description{
Probability represented by a single bar
}
\examples{

barprob(c("Yes", "No"), c(80,20),
        colours=c("green", "red"))
barprob(c("Yes", "Maybe", "No"), c(50,30,20),
        colours=c("green", "orange", "red"))

}
