% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartab.R
\name{bartab}
\alias{bartab}
\title{Bar plot with a table below}
\usage{
bartab(vals, xlab, ylab, vals2, flab, missing = FALSE)
}
\arguments{
\item{vals}{Vector of the x-axis factors.}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{vals2}{(Optional) Vector of the additional factors.}

\item{flab}{(Optional) Labels for the extra factors.}

\item{missing}{Whether a column for missing values is needed.}
}
\value{
GGPlot object
}
\description{
Bar plot with a table below
}
\examples{
data(mtcars)
cyl = factor(mtcars$cyl)
gears = factor(mtcars$gear)
bartab(vals = cyl, xlab = "Cylinders", ylab = "Frequency")
bartab(vals = cyl, xlab = "Cylinders", ylab = "Frequency",
       vals2 = gears, flab = "Gears")
}
