% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninN.R
\name{ninN}
\alias{ninN}
\title{Icons showing n in N people}
\usage{
ninN(
  n,
  N,
  icon,
  additional,
  plt_size = 0.5,
  icon_lib = "RiskCom",
  export,
  colours
)
}
\arguments{
\item{n}{Number with the disease or condition}

\item{N}{Total number of people}

\item{icon}{Path to the icon (file or url)}

\item{additional}{(Optional) Number of additional patients with the risk due to a factor, for example}

\item{plt_size}{Size of the icon}

\item{icon_lib}{Library to fetch the icons from (One of 'file', 'RiskCom', or 'internet')}

\item{export}{(Optional) File to export the image to.}

\item{colours}{Colours for the icons}
}
\value{
GGPlot object
}
\description{
Icons showing n in N people
}
\examples{
# Example showing 2 in 5 people with a disease.
ninN(n=2, N=5, icon="person", plt_size=.3)
# Example with a risk factor increasing the risk from 2 in 5 to 3 in 5
ninN(n=2, N=5, icon="person", additional=1, plt_size=.3)
}
